Teach cosipy to set the datatype used for the inputs to deconvolution.  We can set the type of the dataset contents when they are loaded, but we have to modify the .yml file for deconvolution to control the type of the initial model.

+++ cosipy/image_deconvolution/dataIF_COSI_DC2.py
@@ -88,7 +88,7 @@ class DataIF_COSI_DC2(ImageDeconvolutionDataInterfaceBase):
 
         if isinstance(rsp, FullDetectorResponse):
             logger.info('Loading the response matrix onto your computer memory...')
-            new._load_full_detector_response_on_memory(rsp, is_miniDC2_format)
+            new._load_full_detector_response_on_memory(rsp, is_miniDC2_format, dtype=dtype)
             logger.info('Finished')
         elif isinstance(rsp, Histogram):
             rsp.track_overflow(False)
